+++ nilearn/experimental/surface/_datasets.py
@@ -8,12 +8,12 @@ from __future__ import annotations
 from collections.abc import Sequence
 
 from nilearn import datasets
-from nilearn.experimental.surface import _io
 from nilearn.experimental.surface._surface_image import (
     FileMesh,
     PolyMesh,
     SurfaceImage,
 )
+from nilearn.surface import load_surf_data
 
 
 def load_fsaverage(
@@ -67,10 +67,10 @@ def load_fsaverage_data(
     img = SurfaceImage(
         mesh=fsaverage[mesh_type],
         data={
-            "left": _io.read_array(
+            "left": load_surf_data(
                 old_fsaverage[f"{renaming[data_type]}_left"]
             ),
-            "right": _io.read_array(
+            "right": load_surf_data(
                 old_fsaverage[f"{renaming[data_type]}_right"]
             ),
         },
@@ -104,8 +104,8 @@ def fetch_nki(mesh_type: str = "pial", **kwargs) -> Sequence[SurfaceImage]:
     for left, right in zip(
         nki_dataset["func_left"], nki_dataset["func_right"]
     ):
-        left_data = _io.read_array(left).T
-        right_data = _io.read_array(right).T
+        left_data = load_surf_data(left).T
+        right_data = load_surf_data(right).T
         img = SurfaceImage(
             mesh=fsaverage[mesh_type],
             data={
