+++ nilearn/surface/surface.py
@@ -1237,3 +1237,50 @@ def check_surface(surface):
     surface = load_surface(surface)
     mesh, data = check_mesh_and_data(surface.mesh, surface.data)
     return Surface(mesh, data)
+
+
+def mesh_to_gifti(
+    coordinates,
+    faces,
+    gifti_file,
+):
+    """Write surface mesh to gifti file on disk."""
+    gifti_file = Path(gifti_file)
+    gifti_img = gifti.GiftiImage()
+    coords_array = gifti.GiftiDataArray(
+        coordinates, intent="NIFTI_INTENT_POINTSET", datatype="float32"
+    )
+    faces_array = gifti.GiftiDataArray(
+        faces, intent="NIFTI_INTENT_TRIANGLE", datatype="int32"
+    )
+    gifti_img.add_gifti_data_array(coords_array)
+    gifti_img.add_gifti_data_array(faces_array)
+    gifti_img.to_filename(gifti_file)
+
+
+def data_to_gifti(data, gifti_file):
+    """Save data from Polydata to a gifti file."""
+    # see
+    # https://github.com/nipy/nibabel/blob/master/nibabel/gifti/gifti.py
+    # Only the following are 'supported' for now
+    # - NIFTI_TYPE_UINT8
+    # - NIFTI_TYPE_INT32
+    # - NIFTI_TYPE_FLOAT32
+    if data.dtype in [np.uint16, np.uint32, np.uint64]:
+        data = data.astype(np.uint8)
+    elif data.dtype in [np.int8, np.int16, np.int64]:
+        data = data.astype(np.int32)
+    elif data.dtype in [np.float64]:
+        data = data.astype(np.float32)
+
+    if data.dtype == np.uint8:
+        datatype = "NIFTI_TYPE_UINT8"
+    elif data.dtype == np.int32:
+        datatype = "NIFTI_TYPE_INT32"
+    elif data.dtype == np.float32:
+        datatype = "NIFTI_TYPE_FLOAT32"
+
+    darray = gifti.GiftiDataArray(data=data, datatype=datatype)
+
+    gii = gifti.GiftiImage(darrays=[darray])
+    gii.to_filename(Path(gifti_file))
