+++ nilearn/surface/tests/test_surface.py
@@ -71,6 +71,18 @@ class SurfaceLikeObject:
         return self._data
 
 
+def test_load_surf_data(tmp_path):
+    gifti_img = gifti.GiftiImage()
+    a = np.arange(5)
+    gifti_array = gifti.GiftiDataArray(a, datatype="float32")
+    gifti_img.add_gifti_data_array(gifti_array)
+    gifti_file = tmp_path / "a.gii"
+    gifti_img.to_filename(gifti_file)
+
+    read_a = surface.load_surf_data(gifti_file)
+    assert np.array_equal(a, read_a)
+
+
 def test_load_surf_data_numpy_gt_1pt23():
     """Test loading fsaverage surface.
